Fix `NSPrivacyCollectedDataTypes` array in privacy manifest (#49041)
In https://github.com/flutter/engine/pull/48951 `NSPrivacyCollectedDataTypes` is an array with an empty dictionary, but the privacy report wants an empty array or it errors:

<img width="523" alt="Screenshot 2023-12-14 at 10 02 52â¯AM" src="https://github.com/flutter/flutter/assets/682784/a128b384-db2f-49d7-8b71-7965f2826de8">

When this is fixed the privacy report is blank, but that seems to be an Xcode bug?  Will investigate.

[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style

(cherry picked from commit 0ca2f38ef249948a1e1ebe55f108809ed3bf57ab)